Refactor core components into `Tiles` and `Areas` namespaces
